Made it possible to mock TryCancel() calls.
Needed to mock TryCancel calls for the rdwr stream context
to make the unit tests meaningful.  Then fix both tests and code.

The code was wrong because it was calling the subscribers twice
with the same data.
The tests were wrong because they had setup the wrong mocks in a
couple of places.