Separate the install vs. build CI scripts.
Use a script to create the CI image and a separate script to run
it.  This has almost no effect in the CI builds, but makes it
possible to reproduce CI problems more quickly in my development
workstation.  It also enables separate configurations for the
image, e.g., one where the dependencies are pre-installed.